add util for album

+++ src/main/java/com/example/musicLibrary/services/impl/AlbumServiceImpl.java
@@ -8,10 +8,13 @@ import com.example.musicLibrary.dto.forms.AlbumForm;
 import com.example.musicLibrary.entity.Album;
 import com.example.musicLibrary.entity.Artist;
 import com.example.musicLibrary.services.AlbumService;
+import com.example.musicLibrary.util.AlbumMapper;
+import jakarta.persistence.EntityNotFoundException;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -19,75 +22,85 @@ public class AlbumServiceImpl implements AlbumService {
 
     private final AlbumDaoImpl albumDao;
     private final ArtistDaoImpl artistDao;
-    private final ModelMapper modelMapper;
+    private final AlbumMapper albumMapper;
 
     @Autowired
-    public AlbumServiceImpl(AlbumDaoImpl albumDao, ModelMapper modelMapper, ArtistDaoImpl artistDao) {
+    public AlbumServiceImpl(AlbumDaoImpl albumDao, AlbumMapper albumMapper, ArtistDaoImpl artistDao) {
         this.albumDao = albumDao;
-        this.modelMapper = modelMapper;
         this.artistDao = artistDao;
+        this.albumMapper = albumMapper;
     }
 
     @Override
     public AlbumDTO createAlbum(AlbumDTO albumDTO, long artistId) {
         Artist artist = artistDao.getArtistById(artistId);
+        if (artist == null) {
+            throw new EntityNotFoundException("Artist with id " + artistId + " not found");
+        }
         albumDTO.setArtist(artist);
-        return mapToDTO(albumDao.createAlbum(mapToEntity(albumDTO)));
+        return albumMapper.mapToDTO(albumDao.createAlbum(albumMapper.mapToEntity(albumDTO)));
     }
 
     @Override
     public AlbumDTO getAlbumById(long id) {
-        return mapToDTO(albumDao.getAlbumById(id));
+        if (albumDao.getAlbumById(id) == null)
+            throw new EntityNotFoundException("Album with id " + id + " not found");
+        return albumMapper.mapToDTO(albumDao.getAlbumById(id));
     }
 
     @Override
     public List<AlbumDTO> getAllAlbums() {
         List<Album> albumList = albumDao.getAllAlbums();
-        return albumList.stream().map(this::mapToDTO).toList();
+        if (albumList.isEmpty())
+            throw new EntityNotFoundException("Albums not found");
+        return albumList.stream().map(albumMapper::mapToDTO).toList();
     }
 
 
     @Override
     public AlbumDTO updateAlbum(AlbumForm albumForm) {
         Album albumUpdate = albumDao.getAlbumById(albumForm.getId());
+        if (albumUpdate == null) {
+            throw new EntityNotFoundException("Album with id " + albumForm.getId() + " not found");
+        }
         Artist artist = artistDao.getArtistById(albumForm.getArtistId());
+        if (artist == null) {
+            throw new EntityNotFoundException("Artist with id " + albumForm.getArtistId() + " not found");
+        }
         albumUpdate.setTitle(albumForm.getTitle());
         albumUpdate.setYear(albumForm.getYear());
         albumUpdate.setArtist(artist);
 
         Album newAlbum = albumDao.updateAlbum(albumUpdate);
-        return mapToDTO(newAlbum);
+        return albumMapper.mapToDTO(newAlbum);
     }
 
     @Override
     public void deleteAlbum(long id) {
+        if (albumDao.getAlbumById(id) == null)
+            throw new EntityNotFoundException("Album with id " + id + " not found");
         albumDao.deleteAlbum(id);
     }
 
     @Override
     public List<AlbumDTO> getAlbumsByArtistId(long artistId) {
         List<Album> albumList = albumDao.getAlbumsByArtistId(artistId);
-        return albumList.stream().map(this::mapToDTO).toList();
+        if (albumList.isEmpty())
+            throw new EntityNotFoundException("Albums not found");
+        return albumList.stream().map(albumMapper::mapToDTO).toList();
     }
 
     @Override
     public AlbumDTO getAlbumByArtistIdAndAlbumId(long artistId, long albumId) {
-        return mapToDTO(albumDao.getAlbumByArtistIdAndAlbumId(artistId, albumId));
+        if (albumDao.getAlbumByArtistIdAndAlbumId(artistId, albumId) == null)
+            throw new EntityNotFoundException("Album with id " + albumId + " not found");
+        return albumMapper.mapToDTO(albumDao.getAlbumByArtistIdAndAlbumId(artistId, albumId));
     }
 
     @Override
     public AlbumDTO findAlbumByTitle(String title) {
-        return mapToDTO(albumDao.findAlbumByTitle(title));
-    }
-
-    private Album mapToEntity(AlbumDTO albumDTO) {
-        return modelMapper.map(albumDTO, Album.class);
-    }
-
-    private AlbumDTO mapToDTO(Album newAlbum) {
-        AlbumDTO dto = new AlbumDTO();
-
-        dto.setId(newAlbum.getId());
-        return modelMapper.map(newAlbum, AlbumDTO.class);
+        if (albumDao.findAlbumByTitle(title) == null)
+            throw new EntityNotFoundException("Album with title " + title + " not found");
+        return albumMapper.mapToDTO(albumDao.findAlbumByTitle(title));
     }
 }
